pyproject.toml
[project]
name = "nonebot-plugin-daily-bing"
version = "0.1.0"
description = "NoneBot 每日必应壁纸" 
readme = "README.md"
authors = [
    { name = "lyqgzbl", email = "122811297+lyqgzbl@users.noreply.github.com" }
]
requires-python = ">=3.10, <4.0"
license = "MIT"
dependencies = [
  "httpx >=0.28.0, <1.0.0",
  "nonebot2 >=2.2.1, <3.0.0",
  "aiofiles>=24.1.0",
  "nonebot-plugin-alconna >=0.57.1, <1.0.0",
  "nonebot-plugin-localstore>=0.7.4, <1.0.0",
]

[project.urls]
homepage = "https://github.com/lyqgzbl/nonebot-plugin-daily-bing"
repository = "https://github.com/lyqgzbl/nonebot-plugin-daily-bing"
documentation = "https://github.com/lyqgzbl/nonebot-plugin-daily-bing#readme"

[build-system]
requires = ["uv_build>=0.8.17,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "ruff>=0.14.0",
]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "W", "F", "UP", "C", "T", "PYI", "PT", "Q"]
ignore = ["E402", "C901"]
fixable = ["ALL"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
